Move test_executable to behavioural
